Move code to a different file
