+++ dashboard/src/components/ChartView/HeadingRenderer.tsx
@@ -1,0 +1,20 @@
+import * as React from "react";
+
+// Code from https://github.com/rexxars/react-markdown/issues/69
+function flatten(text: string, child: any): any {
+  return typeof child === "string"
+    ? text + child
+    : React.Children.toArray(child.props.children).reduce(flatten, text);
+}
+
+const HeadingRenderer: React.SFC<{}> = (props: any) => {
+  const children = React.Children.toArray(props.children);
+  const text = children.reduce(flatten, "");
+  const slug = text
+    .toLowerCase()
+    .replace(/[^a-zA-Z0-9_\s]/g, "") // remove punctuation
+    .replace(/\s/g, "-"); // replace spaces with dash
+  return React.createElement("h" + props.level, { id: slug }, props.children);
+};
+
+export default HeadingRenderer;
